会社情報入力画面　（住所検索以外）
